Move character case table declarations from src to detail
