+++ cpp/include/cudf/strings/detail/char_tables.hpp
@@ -35,6 +35,18 @@ using character_flags_table_type = std::uint8_t;
  */
 const character_flags_table_type* get_character_flags_table();
 
+// utilities to dissect a char-table flag
+constexpr uint8_t IS_DECIMAL(uint8_t x) { return ((x) & (1 << 0)); }
+constexpr uint8_t IS_NUMERIC(uint8_t x) { return ((x) & (1 << 1)); }
+constexpr uint8_t IS_DIGIT(uint8_t x) { return ((x) & (1 << 2)); }
+constexpr uint8_t IS_ALPHA(uint8_t x) { return ((x) & (1 << 3)); }
+constexpr uint8_t IS_SPACE(uint8_t x) { return ((x) & (1 << 4)); }
+constexpr uint8_t IS_UPPER(uint8_t x) { return ((x) & (1 << 5)); }
+constexpr uint8_t IS_LOWER(uint8_t x) { return ((x) & (1 << 6)); }
+constexpr uint8_t IS_SPECIAL(uint8_t x) { return ((x) & (1 << 7)); }
+constexpr uint8_t IS_ALPHANUM(uint8_t x) { return ((x) & (0x0F)); }
+constexpr uint8_t IS_UPPER_OR_LOWER(uint8_t x) { return ((x) & ((1 << 5) | (1 << 6))); }
+
 // Type for the character cases table.
 using character_cases_table_type = uint16_t;
 
@@ -51,6 +63,21 @@ using character_cases_table_type = uint16_t;
 const character_cases_table_type* get_character_cases_table();
 
 /**
+ * @brief Case mapping structure for special characters.
+ *
+ * This is used for special mapping of a small set of characters that do not
+ * fit in the character-cases-table.
+ *
+ * @see cpp/src/strings/char_types/char_cases.h
+ */
+struct special_case_mapping {
+  uint16_t num_upper_chars;
+  uint16_t upper[3];
+  uint16_t num_lower_chars;
+  uint16_t lower[3];
+};
+
+/**
  * @brief Returns pointer to device memory that contains the special
  * case mapping table. On first call, this will copy the table into
  * device memory and is guaranteed to be thread-safe.
@@ -61,6 +88,17 @@ const character_cases_table_type* get_character_cases_table();
  * @return Device memory pointer to the special case mapping table
  */
 const struct special_case_mapping* get_special_case_mapping_table();
+
+/**
+ * @brief Get the special mapping table index for a given code-point.
+ *
+ * @see cpp/src/strings/char_types/char_cases.h
+ */
+constexpr uint16_t get_special_case_hash_index(uint32_t code_point)
+{
+  constexpr uint16_t special_case_prime = 499;  // computed from generate_special_mapping_hash_table
+  return static_cast<uint16_t>(code_point % special_case_prime);
+}
 
 }  // namespace detail
 }  // namespace strings
